+++ packages/core/src/components/Checkbox/index.ts
@@ -1,2 +1,3 @@
 export { default } from './Checkbox';
+export { checkboxStyles } from './styles';
 export * from './types';
