+++ packages/widgets/src/components/TableFilters/TableFilterCheckbox.tsx
@@ -1,9 +1,7 @@
-import Checkbox, { ICheckboxProps } from '@components/Checkbox';
-import theme from '@themes/main';
-import { css } from '@emotion/react';
-import { blueInputCheckbox } from '@components/Checkbox/styles';
+import { ICheckboxProps, Checkbox, checkboxStyles } from '@ssa-ui-kit/core';
+import { Theme, css, useTheme } from '@emotion/react';
 
-const getStyles = () => css`
+const getStyles = (theme: Theme) => css`
   margin: 0 0 14px;
   & span {
     margin-left: 16px;
@@ -16,10 +14,11 @@ const getStyles = () => css`
       border-radius: 2px;
     }
   }
-  ${blueInputCheckbox}
+  ${checkboxStyles.blueInput(theme)}
 `;
 
 export const TableFilterCheckbox = (props: ICheckboxProps) => {
-  const styles = getStyles();
+  const theme = useTheme();
+  const styles = getStyles(theme);
   return <Checkbox {...props} css={styles} />;
 };
