+++ packages/widgets/src/components/TableFilters/TableFilterClearButton.tsx
@@ -1,4 +1,4 @@
-import Button from '@components/Button';
+import { Button } from '@ssa-ui-kit/core';
 import { Theme, css, useTheme } from '@emotion/react';
 import { baseButtonStyle } from './styles';
